requirements.txt
openai>=1.50.0
python-dotenv>=1.0.0
chromadb>=0.5.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-timeout>=2.3.0
